+++ amethyst/master/src/assets/formats/mod.rs
@@ -1,0 +1,9 @@
+//! Contains formats that amethyst can load as assets.
+
+pub mod audio;
+
+/// Formats to load textures
+pub mod textures;
+
+/// Formats to load meshes
+pub mod meshes;