+++ amethyst/master/src/ecs/audio/components/mod.rs
@@ -1,0 +1,7 @@
+//! `amethyst` audio ecs components
+
+pub use self::audio_emitter::AudioEmitter;
+pub use self::audio_listener::AudioListener;
+
+mod audio_emitter;
+mod audio_listener;