+++ amethyst/master/src/ecs/audio/mod.rs
@@ -1,0 +1,7 @@
+//! `amethyst` audio ecs module
+
+pub use self::components::*;
+pub use self::systems::*;
+
+pub mod components;
+pub mod systems;