+++ amethyst/master/src/ecs/rendering/components/light.rs
@@ -1,0 +1,12 @@
+//! Light resource handling.
+
+use renderer::prelude::Light;
+
+use ecs::{Component, VecStorage};
+
+/// Wraps `Light` into component
+pub struct LightComponent(pub Light);
+
+impl Component for LightComponent {
+    type Storage = VecStorage<Self>;
+}